--- checkout/rust/payment-customization/default/src/tests.rs
@@ -1,20 +1,0 @@
-use super::*;
-use shopify_function::{run_function_with_input, Result};
-
-#[test]
-fn test_result_contains_no_operations() -> Result<()> {
-    let result = run_function_with_input(
-        function,
-        r#"
-            {
-                "paymentCustomization": {
-                    "metafield": null
-                }
-            }
-        "#,
-    )?;
-    let expected = crate::output::FunctionResult { operations: vec![] };
-
-    assert_eq!(result, expected);
-    Ok(())
-}
